Rearrange stash box client code
